Add links to HF models
