+++ recipes/ESC50/interpret/hparams/piq_vit.yaml
@@ -93,8 +93,8 @@ classifier: !new:speechbrain.lobes.models.ECAPA_TDNN.Classifier
     out_neurons: !ref <out_n_neurons>
     lin_blocks: 1
 
-embedding_model_path: null
-classifier_model_path: null
+embedding_model_path: speechbrain/vit-base-esc50/embedding_model.ckpt
+classifier_model_path: speechbrain/vit-base-esc50/classifier.ckpt
 
 pretrained_esc50: !new:speechbrain.utils.parameter_transfer.Pretrainer
     collect_in: !ref <save_folder>
@@ -125,7 +125,7 @@ compute_istft: !new:speechbrain.processing.features.ISTFT
     hop_length: !ref <hop_length>
     win_length: !ref <win_length>
 
-psi_model: !new:train_piq.VectorQuantizedPSIViT
+psi_model: !new:speechbrain.lobes.models.PIQ.VectorQuantizedPSIFocalNetViT_Audio
     dim: 768
     K: !ref <K>
     shared_keys: 0
