+++ tests/recipes/ESC50.csv
@@ -1,13 +1,13 @@
 Task,Dataset,Script_file,Hparam_file,Data_prep_file,Readme_file,Result_url,HF_repo,test_debug_flags,test_debug_checks,performance
 SoundClassification,ESC50,recipes/ESC50/classification/train.py,recipes/ESC50/classification/hparams/cnn14.yaml,recipes/ESC50/classification/esc50_prepare.py,recipes/ESC50/classification/README.md,https://www.dropbox.com/sh/fbe7l14o3n8f5rw/AACABE1BQGBbX4j6A1dIhBcSa?dl=0,,--data_folder tests/samples/ESC50 --debug --out_n_neurons 3 --use_pretrained 0,"file_exists=[train.py,train_log.txt,log.txt,env.log,hyperparams.yaml,save/label_encoder.txt]",Accuracy=82%
 SoundClassification,ESC50,recipes/ESC50/classification/train.py,recipes/ESC50/classification/hparams/conv2d.yaml,recipes/ESC50/classification/esc50_prepare.py,recipes/ESC50/classification/README.md,https://www.dropbox.com/sh/tl2pbfkreov3z7e/AADwwhxBLw1sKvlSWzp6DMEia?dl=0,,--data_folder tests/samples/ESC50 --debug --out_n_neurons 3 --use_pretrained 0,"file_exists=[train.py,train_log.txt,log.txt,env.log,hyperparams.yaml,save/label_encoder.txt]",Accuracy=75%
-SoundClassification,ESC50,recipes/ESC50/classification/train.py,recipes/ESC50/classification/hparams/focalnet.yaml,recipes/ESC50/classification/esc50_prepare.py,recipes/ESC50/classification/README.md,,,--data_folder tests/samples/ESC50 --debug --out_n_neurons 3,"file_exists=[train.py,train_log.txt,log.txt,env.log,hyperparams.yaml,save/label_encoder.txt]",
-SoundClassification,ESC50,recipes/ESC50/classification/train.py,recipes/ESC50/classification/hparams/vit.yaml,recipes/ESC50/classification/esc50_prepare.py,recipes/ESC50/classification/README.md,,,--data_folder tests/samples/ESC50 --debug --out_n_neurons 3,"file_exists=[train.py,train_log.txt,log.txt,env.log,hyperparams.yaml,save/label_encoder.txt]",
-Interpretability,ESC50,recipes/ESC50/interpret/interpret_amt.py,recipes/ESC50/interpret/hparams/amt_focalnet.yaml,recipes/ESC50/interpret/esc50_prepare.py,recipes/ESC50/interpret/README.md,,,--data_folder tests/samples/ESC50 --debug --use_pretrained 0 --out_n_neurons 3,"file_exists=[train_log.txt,log.txt,interpret_amt.py,env.log,hyperparams.yaml,save/label_encoder.txt]",
-Interpretability,ESC50,recipes/ESC50/interpret/interpret_amt.py,recipes/ESC50/interpret/hparams/amt_vit.yaml,recipes/ESC50/interpret/esc50_prepare.py,recipes/ESC50/interpret/README.md,,,--data_folder tests/samples/ESC50 --debug --use_pretrained 0 --out_n_neurons 3,"file_exists=[train_log.txt,log.txt,interpret_amt.py,env.log,hyperparams.yaml,save/label_encoder.txt]",
+SoundClassification,ESC50,recipes/ESC50/classification/train.py,recipes/ESC50/classification/hparams/focalnet.yaml,recipes/ESC50/classification/esc50_prepare.py,recipes/ESC50/classification/README.md,https://www.dropbox.com/scl/fo/zk101h5xypgi56d777yp5/AGVIfoe56OWInxWf6F57JyQ?rlkey=hmme5c8rnu2sok3jnwbanw7eq&dl=0,,--data_folder tests/samples/ESC50 --debug --out_n_neurons 3,"file_exists=[train.py,train_log.txt,log.txt,env.log,hyperparams.yaml,save/label_encoder.txt]",
+SoundClassification,ESC50,recipes/ESC50/classification/train.py,recipes/ESC50/classification/hparams/vit.yaml,recipes/ESC50/classification/esc50_prepare.py,recipes/ESC50/classification/README.md,https://www.dropbox.com/scl/fo/af59l6mtm0ytqyhz3l7ib/ADGklBYXxil1DWKv5CSMDGk?rlkey=wk5tdh0h26f61e1tn3bh80vys&dl=0,,--data_folder tests/samples/ESC50 --debug --out_n_neurons 3,"file_exists=[train.py,train_log.txt,log.txt,env.log,hyperparams.yaml,save/label_encoder.txt]",
+Interpretability,ESC50,recipes/ESC50/interpret/interpret_amt.py,recipes/ESC50/interpret/hparams/amt_focalnet.yaml,recipes/ESC50/interpret/esc50_prepare.py,recipes/ESC50/interpret/README.md,https://www.dropbox.com/scl/fo/0hheboei1b35mlrhwj6mt/AOeCdNstN3h8UqFxv0abT7M?rlkey=kx0d1t5v5hqawqwr5ir9weihq&dl=0,,--data_folder tests/samples/ESC50 --debug --use_pretrained 0 --out_n_neurons 3,"file_exists=[train_log.txt,log.txt,interpret_amt.py,env.log,hyperparams.yaml,save/label_encoder.txt]",
+Interpretability,ESC50,recipes/ESC50/interpret/interpret_amt.py,recipes/ESC50/interpret/hparams/amt_vit.yaml,recipes/ESC50/interpret/esc50_prepare.py,recipes/ESC50/interpret/README.md,https://www.dropbox.com/scl/fo/vlluiqiirlprl3oa7h4sj/APrEFgcIiWjdQhDUEZuNook?rlkey=bhswfspzklypu7k8ndh8lm3st&dl=0,,--data_folder tests/samples/ESC50 --debug --use_pretrained 0 --out_n_neurons 3,"file_exists=[train_log.txt,log.txt,interpret_amt.py,env.log,hyperparams.yaml,save/label_encoder.txt]",
 Interpretability,ESC50,recipes/ESC50/interpret/train_l2i.py,recipes/ESC50/interpret/hparams/l2i_cnn14.yaml,recipes/ESC50/interpret/esc50_prepare.py,recipes/ESC50/interpret/README.md,https://www.dropbox.com/sh/cli2gm8nb4bthow/AAAKnzU0c80s_Rm7wx4i_Orza?dl=0,,--data_folder tests/samples/ESC50 --debug --use_pretrained 0,"file_exists=[train_l2i.py,train_log.txt,log.txt,env.log,hyperparams.yaml,save/label_encoder.txt]",
 Interpretability,ESC50,recipes/ESC50/interpret/train_l2i.py,recipes/ESC50/interpret/hparams/l2i_conv2d.yaml,recipes/ESC50/interpret/esc50_prepare.py,recipes/ESC50/interpret/README.md,https://www.dropbox.com/sh/gcpk9jye9ka08n0/AAB-m10r1YEH0rJdUMrCwizUa?dl=0,,--data_folder tests/samples/ESC50 --debug --use_pretrained 0,"file_exists=[train_l2i.py,train_log.txt,log.txt,env.log,hyperparams.yaml,save/label_encoder.txt]",
 Interpretability,ESC50,recipes/ESC50/interpret/train_nmf.py,recipes/ESC50/interpret/hparams/nmf.yaml,recipes/ESC50/interpret/esc50_prepare.py,recipes/ESC50/interpret/README.md,https://www.dropbox.com/sh/01exv8dt3k6l1kk/AADuKmikAPwMw5wlulojd5Ira?dl=0,,--data_folder tests/samples/ESC50 --debug,"file_exists=[train_log.txt,train_nmf.py,log.txt,env.log,hyperparams.yaml,save/label_encoder.txt]",
 Interpretability,ESC50,recipes/ESC50/interpret/train_piq.py,recipes/ESC50/interpret/hparams/piq.yaml,recipes/ESC50/interpret/esc50_prepare.py,recipes/ESC50/interpret/README.md,https://www.dropbox.com/sh/v1x5ks9t67ftysp/AABo494rDElHTiTpKR_6PP_ua?dl=0,,--data_folder tests/samples/ESC50 --debug --use_pretrained 0 --out_n_neurons 3 --number_of_epochs 1,"file_exists=[train_log.txt,log.txt,train_piq.py,env.log,hyperparams.yaml,save/label_encoder.txt]",
-Interpretability,ESC50,recipes/ESC50/interpret/train_piq.py,recipes/ESC50/interpret/hparams/piq_focalnet.yaml,recipes/ESC50/interpret/esc50_prepare.py,recipes/ESC50/interpret/README.md,,,--data_folder tests/samples/ESC50 --debug --use_pretrained 0 --out_n_neurons 2 --number_of_epochs 1 --K 1,"file_exists=[train_log.txt,log.txt,train_piq.py,env.log,hyperparams.yaml,save/label_encoder.txt]",
-Interpretability,ESC50,recipes/ESC50/interpret/train_piq.py,recipes/ESC50/interpret/hparams/piq_vit.yaml,recipes/ESC50/interpret/esc50_prepare.py,recipes/ESC50/interpret/README.md,,,--data_folder tests/samples/ESC50 --debug --use_pretrained 0 --out_n_neurons 2 --number_of_epochs 1 --K 1,"file_exists=[train_log.txt,log.txt,train_piq.py,env.log,hyperparams.yaml,save/label_encoder.txt]",
+Interpretability,ESC50,recipes/ESC50/interpret/train_piq.py,recipes/ESC50/interpret/hparams/piq_focalnet.yaml,recipes/ESC50/interpret/esc50_prepare.py,recipes/ESC50/interpret/README.md,https://www.dropbox.com/scl/fo/6mvxb32f0g1i8b4lkdjoq/AGD1xNF8Of2_IXeEsbpXtQE?rlkey=llefue4rxalqyqwxqtwrn8qii&dl=0,,--data_folder tests/samples/ESC50 --debug --use_pretrained 0 --out_n_neurons 2 --number_of_epochs 1 --K 1,"file_exists=[train_log.txt,log.txt,train_piq.py,env.log,hyperparams.yaml,save/label_encoder.txt]",
+Interpretability,ESC50,recipes/ESC50/interpret/train_piq.py,recipes/ESC50/interpret/hparams/piq_vit.yaml,recipes/ESC50/interpret/esc50_prepare.py,recipes/ESC50/interpret/README.md,https://www.dropbox.com/scl/fo/nz4lqwumgz03nanmf9xai/AI21fGwSOzsVvyegTJUEtz4?rlkey=40yjchqgkhcrhbxsa30m3rr6w&dl=0,,--data_folder tests/samples/ESC50 --debug --use_pretrained 0 --out_n_neurons 2 --number_of_epochs 1 --K 1,"file_exists=[train_log.txt,log.txt,train_piq.py,env.log,hyperparams.yaml,save/label_encoder.txt]",
